requirements.txt
# Web Framework
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23

# Async and HTTP
aiohttp>=3.8.0
asyncio>=3.4.3
playwright>=1.20.0

# HTML Parsing
beautifulsoup4>=4.9.3
lxml>=4.9.3

# System monitoring
psutil>=5.8.0

# Security
PyYAML>=5.4.1
cryptography>=41.0.7
colorama>=0.4.6

# Utilities
python-dotenv>=1.0.0
urllib3>=1.26.5

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Additional requirements
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
colorama>=0.4.4
psutil>=5.8.0
urllib3>=1.26.7
python-dotenv>=0.19.0
pydantic>=1.9.0